encapsulate Node and Participant.
Fix nasty FF mtls bug - D'oh. That was a killer.

remove thread.sleeps. include initial inserts on choam test genesis.
