+++ sql-state/src/test/java/com/salesforce/apollo/state/CHOAMTest.java
@@ -71,17 +71,19 @@ import com.salesforce.apollo.utils.Utils;
 public class CHOAMTest {
     private static final int CARDINALITY = 5;
 
-    private static final List<Transaction> GENESIS_DATA    = CHOAM.toGenesisData(MigrationTest.initializeBookSchema());
+    private static final List<Transaction> GENESIS_DATA;
     private static final Digest            GENESIS_VIEW_ID = DigestAlgorithm.DEFAULT.digest("Give me food or give me slack or kill me".getBytes());
     private static final boolean           LARGE_TESTS     = Boolean.getBoolean("large_tests");
     static {
         Thread.setDefaultUncaughtExceptionHandler((t, e) -> {
             LoggerFactory.getLogger(CHOAMTest.class).error("Error on thread: {}", t.getName(), e);
         });
+        var txns = MigrationTest.initializeBookSchema();
+        txns.add(initialInsert());
+        GENESIS_DATA = CHOAM.toGenesisData(txns);
     }
 
     private File                               baseDir;
-    private Map<Digest, List<Digest>>          blocks;
     private File                               checkpointDirBase;
     private Map<Digest, CHOAM>                 choams;
     private List<SigningMember>                members;
@@ -131,7 +133,6 @@ public class CHOAMTest {
         baseDir = new File(System.getProperty("user.dir"), "target/cluster-" + Utils.bitStreamEntropy().nextLong());
         Utils.clean(baseDir);
         baseDir.mkdirs();
-        blocks = new ConcurrentHashMap<>();
         Random entropy = new Random();
         var context = new ContextImpl<>(DigestAlgorithm.DEFAULT.getOrigin(), CARDINALITY, 0.2, 3);
         var metrics = new ChoamMetricsImpl(context.getId(), registry);
@@ -182,13 +183,6 @@ public class CHOAMTest {
         System.out.println("Warm up");
         routers.values().forEach(r -> r.start());
         choams.values().forEach(ch -> ch.start());
-
-        Thread.sleep(2_000);
-
-        final var initial = choams.get(members.get(0).getId())
-                                  .getSession()
-                                  .submit(ForkJoinPool.commonPool(), initialInsert(), timeout, txScheduler);
-        initial.get(timeout.toMillis(), TimeUnit.MILLISECONDS);
 
         for (int i = 0; i < clientCount; i++) {
             updaters.entrySet().stream().map(e -> {
@@ -279,8 +273,6 @@ public class CHOAMTest {
 
                                                            @Override
                                                            public void beginBlock(ULong height, Digest hash) {
-                                                               blocks.computeIfAbsent(m.getId(), k -> new ArrayList<>())
-                                                                     .add(hash);
                                                                up.getExecutor().beginBlock(height, hash);
                                                            }
 
@@ -298,15 +290,13 @@ public class CHOAMTest {
                                                            @Override
                                                            public void genesis(Digest hash,
                                                                                List<Transaction> initialization) {
-                                                               blocks.computeIfAbsent(m.getId(), k -> new ArrayList<>())
-                                                                     .add(hash);
                                                                up.getExecutor().genesis(hash, initialization);
                                                            }
                                                        })
                                                        .build()));
     }
 
-    private Txn initialInsert() {
+    private static Txn initialInsert() {
         return Txn.newBuilder()
                   .setBatch(batch("insert into books values (1001, 'Java for dummies', 'Tan Ah Teck', 11.11, 11)",
                                   "insert into books values (1002, 'More Java for dummies', 'Tan Ah Teck', 22.22, 22)",
